requirements.txt
# Common
python-dotenv==0.21.0
requests==2.28.1

# Django Libs
django==3.2
gunicorn==20.1.0

# Celery
celery==5.2.7
django-celery-results==2.4.0

redis==4.3.4
async-timeout==4.0.2
deprecated==1.2.13
packaging==21.3
pyparsing==3.0.9
wrapt==1.14.1


# Telegram
python-telegram-bot==13.14